Update Modified working code
This is a modified working version of the original script published in Remote Sensing. GEE made some changes to the Landsat 5 dataset that rendered previous code unusable. The changes can be summarised in:
a.  Use of the atmospherically corrected surface reflectance instead of TOA.
b.  Inclussion of a bit cloud mask instead of the Fmask algortihm used in the previous TOA dataset.